refactor(bytes, immutable-arraybuffer, eslint-plugin): split installs, rename symbols, move freezable pony into @endo/bytes (#417 review 4424448137)
Implements maintainer review feedback on PR #417:

- Split monolithic `packages/bytes/src/spackle-install.js` into
  per-operation install modules (one file per source module's
  rendezvous): `install-to-immutable.js`,
  `install-transfer-to-immutable.js`, `install-concat-immutables.js`,
  `install-from-immutable.js`, `install-from-string.js`,
  `install-to-string.js`, and `install-freezable-typedarrays.js`,
  plus a shared `install-helpers.js` that captures the intrinsics
  once and exports the `installOrAdopt` helper. Consumers of an
  individual operation no longer load the entire install bundle.

- Drop Hungarian `Function` suffixes on exported references
  (`bytesFromTextFunction` -> `installedFromTextValue`, etc.) and
  simplify the registered-symbol names:
  `Symbol.for('sliceBufferToImmutable')` ->
  `Symbol.for('sliceToImmutable')` (same name as the proposed
  `ArrayBuffer.prototype.sliceToImmutable` method, so a native
  implementation lands at the same rendezvous),
  `Symbol.for('transferBufferToImmutable')` ->
  `Symbol.for('transferToImmutable')`,
  `Symbol.for('bytesFromImmutable')` -> `Symbol.for('fromImmutable')`,
  `Symbol.for('toUtf8String')` -> `Symbol.for('toText')`,
  `Symbol.for('fromUtf8String')` -> `Symbol.for('fromText')`.

- Add `Symbol.for('toStrictText')` paired with `toText` for the
  fatal-throw UTF-8 decode variant; the lenient `toText` substitutes
  U+FFFD on invalid input, the strict variant throws.

- Sweep the word "spackle" from code identifiers, file names, and
  test variable names; design-document framing of the architectural
  pattern (immutable-arraybuffer DESIGN.md, README.md) retains the
  term. Follows the `@endo/harden` naming precedent: the registered
  symbol bears the operation's own name, not a package-pattern
  prefix.

- Move `freezable-typedarray-pony.js` and its test from
  `@endo/immutable-arraybuffer` into `@endo/bytes` (internal to
  bytes, not exported). To preserve the encapsulation of
  `hiddenBuffers` and `reverseHiddenBuffers`, the move introduces
  a narrow `@endo/immutable-arraybuffer/private-for-bytes.js`
  subpath that re-exports only what the freezable pony needs;
  the subpath is documented as bytes-internal and not advertised
  in the README.

- Update the eslint-plugin's `no-direct-codec-or-typedarray-constructor`
  rule allow-list to reference the new file paths
  (`install-helpers.js` and `freezable-typedarray-pony.js` under
  bytes) and drop "spackle" from suggestion messages.

- Sweep DESIGN.md and README.md to point to the moved file's new
  home in `@endo/bytes`.

Coverage rises to 97.1% on the bytes package (was already high).
All 55 bytes tests and 37 immutable-arraybuffer tests still pass.
