+++ packages/bytes/src/concat.js
@@ -1,14 +1,14 @@
 import harden from '@endo/harden';
-import { Uint8ArrayCaptured } from './spackle-install.js';
+import { CapturedUint8Array } from './install-helpers.js';
 
 /**
  * Concatenates a list of `Uint8Array` chunks into a single contiguous
  * `Uint8Array`.
  * Empty input yields an empty `Uint8Array`.
  *
- * Uses the realm's `Uint8Array` constructor as captured by the spackle
- * at module load (see `./spackle-install.js`); does not rely on the
- * post-lockdown `globalThis.Uint8Array` binding being unmodified.
+ * Uses the realm's `Uint8Array` constructor as captured at module load
+ * (see `./install-helpers.js`); does not rely on the post-lockdown
+ * `globalThis.Uint8Array` binding being unmodified.
  *
  * @param {readonly Uint8Array[]} chunks
  * @returns {Uint8Array}
@@ -18,7 +18,7 @@ export const concatBytes = chunks => {
   for (const chunk of chunks) {
     totalLength += chunk.length;
   }
-  const result = new Uint8ArrayCaptured(totalLength);
+  const result = new CapturedUint8Array(totalLength);
   let offset = 0;
   for (const chunk of chunks) {
     result.set(chunk, offset);
